make front

+++ server/src/main/java/com/example/lgcns/security/JwtConstant.java
@@ -9,7 +9,7 @@ public final class JwtConstant {
 
 	public static final String TOKEN_HEADER = "Authorization";
 	public static final String TOKEN_PREFIX = "Bearer ";
-	public static final long TOKEN_EXPIRE = 864000000; // 10 days
+	public static final int TOKEN_EXPIRE = 864000000; // 10 days
 
 	public static final String TOKEN_TYPE = "JWT";
 	public static final String TOKEN_ISSUER = "secure-api";
